requirements.txt
Flask==2.2.2
jinja2==3.1.2
torch==2.1.0 # You can specify whichever version you need
torchvision==0.16.0
Werkzeug==2.2.2
pandas==2.1.1
numpy==1.26.0
faiss-cpu==1.7.4
sentencepiece==0.1.96
